+++ src/background-chrome.ts
@@ -1,30 +1,15 @@
 /**
  * This is the background script that runs independent of any document. It
- * listens to main frame requests and kicks in if the headers indicate JSON. If
- * we have the filterResponseData API available, we will use that to change the
- * page to what Chrome displays for JSON (this is only used in Firefox). Then a
- * content script reformats the page.
+ * listens to main frame requests and records if the headers indicate the URL is
+ * JSON. When the content script runs, it can ask this script if the page is
+ * JSON, and if so, it will format the page.
  */
 
-import { isJSONContentType } from "./content-type";
-
-function isRedirect(status: number) {
-  return status >= 300 && status < 400;
-}
+import { addJsonUrl, installMessageListener, isEventJSON } from "./background-common";
 
 function detectJSON(event: chrome.webRequest.WebResponseHeadersDetails) {
-  if (!event.responseHeaders || event.type !== "main_frame" || isRedirect(event.statusCode)) {
-    return;
-  }
-  for (const header of event.responseHeaders) {
-    if (
-      header.name.toLowerCase() === "content-type" &&
-      header.value &&
-      isJSONContentType(header.value)
-    ) {
-      addJsonUrl(event.url);
-      break;
-    }
+  if (isEventJSON(event)) {
+    addJsonUrl(event.url);
   }
 
   return { responseHeaders: event.responseHeaders };
@@ -37,35 +22,4 @@ chrome.webRequest.onHeadersReceived.addListener(
   ["responseHeaders"],
 );
 
-// Listen for a message from the content script to decide whether to operate on
-// the page. Calls sendResponse with a boolean that's true if the content script
-// should run, and false otherwise.
-chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
-  if (message !== "jsonview-is-json") {
-    return;
-  }
-
-  if (!sender.url) {
-    sendResponse(false);
-    return;
-  }
-
-  if (sender.url.startsWith("file://") && sender.url.endsWith(".json")) {
-    sendResponse(true);
-    return;
-  }
-
-  hasJsonUrl(sender.url).then(sendResponse);
-  return true; // this means "we're going to sendResponse asynchronously"
-});
-
-async function addJsonUrl(url: string) {
-  await chrome.storage.session.set({ [url]: true });
-}
-
-async function hasJsonUrl(url: string) {
-  const stored = await chrome.storage.session.get(url);
-  const present = url in stored;
-  await chrome.storage.session.remove(url);
-  return present;
-}
+installMessageListener();
